backend/requirements.txt
django>=5.1,<5.2
djangorestframework>=3.15,<4.0
django-cors-headers>=4.4,<5.0
psycopg[binary]>=3.2,<4.0
python-dotenv>=1.0,<2.0
retell-sdk>=4.0,<5.0
gunicorn>=23.0,<24.0
razorpay>=1.4,<2.0
